refactor: @types 폴더 src 외부로 이동
