+++ @types/api/commons.d.ts
@@ -1,0 +1,8 @@
+declare module "@pandamarket-api" {
+  type QueryString = string | string[];
+
+  interface DataFormat<T> {
+    list: T[];
+    totalCount: number;
+  }
+}
