--- src/@types/device.ts
@@ -1,9 +1,0 @@
-const Device = {
-  Mobile: "mobile",
-  Tablet: "tablet",
-  PC: "pc",
-} as const;
-
-export default Device;
-
-export type DeviceValue = (typeof Device)[keyof typeof Device];
