+++ src/variables/Device.ts
@@ -1,0 +1,7 @@
+enum Device {
+  Mobile = "mobile",
+  Tablet = "tablet",
+  PC = "pc",
+}
+
+export default Device;
