+++ src/variables/SortType.ts
@@ -1,0 +1,9 @@
+export enum SortTypeAtUI {
+  Recent = "최신순",
+  Like = "좋아요순",
+}
+
+export enum SortTypeAtAPI {
+  Recent = "recent",
+  Like = "like",
+}
