+++ tsconfig.json
@@ -24,5 +24,6 @@
     "next.config.js",
     "postcss.config.js"
   ],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules"],
+  "typeRoots": ["./node_modules/@types", "./@types"]
 }
